Move Calculator class into its own directory
